// functions/src/models/userResumeSection.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SectionPhrases } from './sectionPhrases';
import { UserSectionFields } from './userSectionFields';

export interface UserResumeSection { 
    sectionID?: any;
    resumeID?: any;
    resumeSectionID?: string;
    sectionIndex?: number;
    sectionName?: string;
    overrideSectionName?: string;
    description?: string;
    sectionIconRef?: string;
    overrideSectionIconRef?: string;
    helpIconRef?: string;
    overrideHelpIconRef?: string;
    deleteIconRef?: string;
    overrideDeleteIconRef?: string;
    associatedCSS?: string;
    overrideAssociatedCSS?: string;
    sectionBarColor?: string;
    overrideSectionBarColor?: string;
    sectionPhrases?: Array<SectionPhrases>;
    sectionFields?: Array<UserSectionFields>;
}