// functions/src/models/userSectionFields.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserSectionFields { 
    sectionFieldID?: string;
    fieldID?: any;
    resumeSectionID?: any;
    fieldName?: string;
    fieldLabel?: string;
    overrideFieldLabel?: string;
    fieldLabelFont?: string;
    overrideFieldLabelFont?: string;
    fieldLabelFontSize?: string;
    overrideFieldLabelFontSize?: string;
    fieldLabelFormatting?: string;
    overrideFieldLabelFormatting?: string;
    fieldValueText?: string;
    fieldValuePictureRef?: string;
    fieldTextFont?: string;
    overrideFieldTextFont?: string;
    fieldTextFontSize?: string;
    overrideFieldTextFontSize?: string;
    fieldTextFormatting?: string;
    overrideFieldTextFormatting?: string;
    associatedCSS?: string;
    overrideAssociatedCSS?: string;
}