// functions/src/models/userResumeRequest.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UserResumeRequest { 
    requestID?: string;
    UserID?: string;
    templateID?: string;
    requestDateTime?: string;
    assignedToUser?: string;
    assignmentHistory?: string;
    completionDateTime?: string;
    requestStatus?: string;
}